// ui/src/generated/types/groupPayloadClass.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Smartphone Test Farm
 * Control and manages real Smartphone devices from browser and restful apis
 * OpenAPI spec version: 2.4.3
 */

/**
 * Group class; privileged value => debug, bookable, standard
 */
export type GroupPayloadClass = (typeof GroupPayloadClass)[keyof typeof GroupPayloadClass]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GroupPayloadClass = {
  once: 'once',
  bookable: 'bookable',
  hourly: 'hourly',
  daily: 'daily',
  weekly: 'weekly',
  monthly: 'monthly',
  quaterly: 'quaterly',
  halfyearly: 'halfyearly',
  yearly: 'yearly',
  debug: 'debug',
  standard: 'standard',
} as const